+++ module/handlebars.ts
@@ -1,5 +1,38 @@
 import { Helpers } from './helpers';
 
+export const preloadHandlebarsTemplates = async () => {
+    const templatePaths = [
+        'systems/shadowrun5e/templates/actor/parts/actor-equipment.html',
+        'systems/shadowrun5e/templates/actor/parts/actor-spellbook.html',
+        'systems/shadowrun5e/templates/actor/parts/actor-skills.html',
+        'systems/shadowrun5e/templates/actor/parts/actor-matrix.html',
+        'systems/shadowrun5e/templates/actor/parts/actor-actions.html',
+        'systems/shadowrun5e/templates/actor/parts/actor-config.html',
+        'systems/shadowrun5e/templates/actor/parts/actor-bio.html',
+        'systems/shadowrun5e/templates/actor/parts/actor-social.html',
+        'systems/shadowrun5e/templates/item/parts/description.html',
+        'systems/shadowrun5e/templates/item/parts/technology.html',
+        'systems/shadowrun5e/templates/item/parts/header.html',
+        'systems/shadowrun5e/templates/item/parts/weapon-ammo-list.html',
+        'systems/shadowrun5e/templates/item/parts/weapon-mods-list.html',
+        'systems/shadowrun5e/templates/item/parts/action.html',
+        'systems/shadowrun5e/templates/item/parts/damage.html',
+        'systems/shadowrun5e/templates/item/parts/opposed.html',
+        'systems/shadowrun5e/templates/item/parts/spell.html',
+        'systems/shadowrun5e/templates/item/parts/complex_form.html',
+        'systems/shadowrun5e/templates/item/parts/weapon.html',
+        'systems/shadowrun5e/templates/item/parts/armor.html',
+        'systems/shadowrun5e/templates/item/parts/matrix.html',
+        'systems/shadowrun5e/templates/item/parts/sin.html',
+        'systems/shadowrun5e/templates/item/parts/contact.html',
+        'systems/shadowrun5e/templates/item/parts/lifestyle.html',
+        'systems/shadowrun5e/templates/item/parts/ammo.html',
+        'systems/shadowrun5e/templates/item/parts/modification.html',
+    ];
+
+    return loadTemplates(templatePaths);
+};
+
 export const registerHandlebarHelpers = () => {
     Handlebars.registerHelper('localizeOb', function (strId, obj) {
         if (obj) strId = obj[strId];
@@ -26,17 +59,17 @@ export const registerHandlebarHelpers = () => {
         if (arr.includes(val)) return options.fn(this);
         else return options.inverse(this);
     });
-// if greater than
+    // if greater than
     Handlebars.registerHelper('ifgt', function (v1, v2, options) {
         if (v1 > v2) return options.fn(this);
         else return options.inverse(this);
     });
-// if not equal
+    // if not equal
     Handlebars.registerHelper('ifne', function (v1, v2, options) {
         if (v1 !== v2) return options.fn(this);
         else return options.inverse(this);
     });
-// if equal
+    // if equal
     Handlebars.registerHelper('ife', function (v1, v2, options) {
         if (v1 === v2) return options.fn(this);
         else return options.inverse(this);
@@ -84,5 +117,4 @@ export const registerHandlebarHelpers = () => {
         }
         return icon;
     });
-
-}
+};
